Add resultFileMinClasspathCount flag
When writing the result file, it becomes really large for big projects
because every entry on the classpath is listed out in the file. This
flag sets a minimum count of classpath occurences for the files
written into the result file. Default setting is 2 occurences so any
file or class that is unique on the class path will not be listed.